==> 01_metadata_cleaning.Rmd <==
---
title: "01_metadata_cleaning"
author: "Jasmine"
date: "12/02/2020"
output: html_document
---

# Data cleaning
_Attempts to standardize columns into the same format and seperate out depth and country information_

## dependancies
```{r dependancies}
library(tidyverse)
library(leaflet)
library(readxl)
library(janitor)
library(googlesheets4)
library(ptm)
```

### Load file into R
- don't worry about the coerced from NA warning messages
- it is the issue from 
```{r getting data from google sheets}
# need to authenticate (see popup window)
#ptm <- masterlist()

pml <- "https://docs.google.com/spreadsheets/d/1vzVIT5gjQ0yCGwbAyqB4f8pltOkLfWz27TIFTxnm9hk/edit#gid=0"
ptm <- googlesheets4::read_sheet(pml, sheet = "All PTM data", skip = 1,
                            col_types = "iccccccDcccccccccccccccccccccccccc-iiccc")
```

```{r coodinate cleaning}
# all records that have decimal degrees
ptm_na <- filter(
  ptm,
  !is.na(Latitude)) %>% 
  mutate(Latitude = as.character(Latitude),
         Longitude = as.character(Longitude))

ptm_ll <- ptm %>% 
  filter(!is.na(Latitude),
         !str_detect(Latitude, "\""),
         !str_detect(Latitude, "\'"),
         !str_detect(Latitude, "N")) %>%
  mutate(Latitude = as.numeric(Latitude),
         Longitude = as.numeric(Longitude))
```

## Mapping the data in leaflet reveals the issues
- coordinates are poor for Chile (- are not in latitude or longitude) -> plotted in Kyrzkstan
### plotting data to check work
```{r initial map echo=FALSE}
leaflet() %>%
  addProviderTiles("CartoDB") %>%
  addCircleMarkers(
    lng = ptm_ll$Longitude, lat = ptm_ll$Latitude,
    popup = paste(
      ptm_ll$Locality, ptm_ll$`PTM#`,
      "Lat", ptm_ll$Latitude, "Long", ptm_ll$Longitude
    ),
    clusterOptions = markerClusterOptions()
  )
```

# GPS Coordinates
- Cleaning up all that are not decimal degrees
- non decimal degree coordinates are less machine readable (ie DD"MM'SS format may mutate if the file is saved differently)

### transforming data to match
```{r fixing names}
ptm_ll$Longitude <- ifelse(grepl("Chile", ptm_ll$Locality) & !grepl("-", ptm_ll$Latitude),
  paste0("-", ptm_ll$Longitude),
  ptm_ll$Longitude
)

ptm_ll$Longitude <- ifelse(grepl("[Ii]sland", ptm_ll$Locality) & !grepl("-", ptm_ll$Longitude),
  paste0("-", ptm_ll$Longitude),
  ptm_ll$Longitude
)

ptm_ll$Latitude <- ifelse(grepl("Chile", ptm_ll$Locality) & !grepl("-", ptm_ll$Latitude),
  paste0("-", ptm_ll$Latitude),
  ptm_ll$Latitude
)
```

convert to decimal degrees
- need to resolve: 1718 (in NT?)
- converts notation (ie. " ') to spaces to be passed to conv_unit (format accepted)
- careful! needed to add negative sign to make sure W values are properly converted
```{r cleaning coordinates}
# all records that need cleaning
xls_issue <- filter(
  ptm,
  !is.na(Latitude),
  !ptm$`PTM#` %in% ptm_ll$`PTM#`)

# change the symbols to a space
xls_issue$Latitude <- gsub("°", " ", xls_issue$Latitude)
xls_issue$Latitude <- gsub("'", " ", xls_issue$Latitude)
xls_issue$Latitude <- gsub('\"', "", xls_issue$Latitude)
xls_issue$Latitude <- gsub("N", "", xls_issue$Latitude)
xls_issue$Latitude <- trimws(xls_issue$Latitude, which = "both")

xls_issue$Longitude <- gsub("°", " ", xls_issue$Longitude)
xls_issue$Longitude <- ifelse(grepl("W", xls_issue$Longitude),
  paste0("-", xls_issue$Longitude),
  xls_issue$Longitude
)
xls_issue$Longitude <- gsub("W", "", xls_issue$Longitude)
xls_issue$Longitude <- gsub("E", "", xls_issue$Longitude)
xls_issue$Longitude <- gsub("'", " ", xls_issue$Longitude)
xls_issue$Longitude <- gsub('\"', "", xls_issue$Longitude)
xls_issue$Longitude <- trimws(xls_issue$Longitude, which = "both")

# convert from decimal minutes to decimal degrees
xls_issue$Latitude <- measurements::conv_unit(xls_issue$Latitude, from = "deg_min_sec", to = "dec_deg")
xls_issue$Longitude <- measurements::conv_unit(xls_issue$Longitude, from = "deg_min_sec", to = "dec_deg")
```

putting back the data together and plotting
```{r final map}
xls_na <- filter(
  ptm,
  is.na(Latitude)
)

ptm_gps <- rbind(ptm_ll, xls_issue, xls_na)

ptm_gps <- ptm_gps %>%
  mutate(
    Latitude = as.numeric(Latitude),
    Longitude = as.numeric(Longitude)
  )
```
## fixed map
```{r}
leaflet() %>%
  addProviderTiles("CartoDB") %>%
  addCircleMarkers(
    lng = ptm_gps$Longitude, lat = ptm_gps$Latitude,
    popup = paste(
      ptm_gps$Locality, ptm_ll$`PTM#`,
      "Lat", ptm_gps$Latitude,
      "Long", ptm_gps$Longitude
    ),
    clusterOptions = markerClusterOptions()
  )
```

## Final Determination
- cleaning up names
```{r determination}
species <- ptm_gps %>% count(`Final determination`, sort = TRUE)

ptm_gps$`Final determination` <- gsub("Chondrachanthus", "Chondracanthus", ptm_gps$`Final determination`)
ptm_gps$`Final determination` <- gsub("\r\n", "", ptm_gps$`Final determination`)
ptm_gps$`Final determination` <- gsub("Sp\\.", "sp.", ptm_gps$`Final determination`)
ptm_gps$`Final determination` <- gsub("sp. 2", "sp.2", ptm_gps$`Final determination`)
ptm_gps$`Final determination` <- gsub("unknown species", "unknown", ptm_gps$`Final determination`)
```

## Coralline/ Red/ Green / Brown
ensuring uniform data
```{r eval=FALSE, include=FALSE}
group_by(ptm_gps, `Red/Coralline/Green/Brown`) %>%
  summarise()

ptm_gps$`Red/Coralline/Green/Brown` <- gsub("red", "Red", ptm_gps$`Red/Coralline/Green/Brown`)
ptm_gps$`Red/Coralline/Green/Brown` <- gsub("\r\n", "", ptm_gps$`Red/Coralline/Green/Brown`)
```

## correcting for inconsistencies
### Primary Collector and Determiners
using gsub to fix
```{r names}
group_by(ptm_gps, `Primary Collector`) %>%
  summarise()

ptm_gps$`Primary Collector` <- gsub("P.T.Martone", "Patrick T. Martone", ptm_gps$`Primary Collector`)
ptm_gps$`Primary Collector` <- gsub("PTMartone", "Patrick T. Martone", ptm_gps$`Primary Collector`)
ptm_gps$`Primary Collector` <- gsub("Samuel", "Sam", ptm_gps$`Primary Collector`)
ptm_gps$`Primary Collector` <- gsub("Vanmaanen", "VanMaanen", ptm_gps$`Primary Collector`)

group_by(ptm_gps, `Other collectors`) %>%
  summarise()

group_by(ptm_gps, `Determined by`) %>%
  summarise()

ptm_gps$`Determined by` <- gsub("Samuel", "Sam", ptm_gps$`Determined by`)

group_by(ptm_gps, `Fragments available?`) %>%
  summarise()

ptm_gps$`Fragments available?` <- gsub("yes", "Yes", ptm_gps$`Fragments available?`)
ptm_gps$`Fragments available?` <- gsub("y", "Yes", ptm_gps$`Fragments available?`)
```

## Fixing depth
```{r}
ptm_gps$Habitat <- gsub("Subtidal", "subtidal", ptm_gps$Habitat)

ptm_d <- ptm_gps %>%
  # separate(col = Habitat, into = c("Habitat", "d", "Other"), sep = ",") %>%
  mutate(
    dep = str_extract(Habitat, "(\\d{0,2}\\+?\\-?\\d{1,2}\\.\\w{1,2})|(\\d{0,2}\\+?\\-?\\d{1,2}\\s*\\w{1,2})"),
    Depth = if_else(!is.na(dep),
      dep,
      Depth
    )
  ) %>% 
  select(-dep)

```

## Date collected
- wasn't able to get everything quite the right format but at least better than before
```{r updating dates}
# DD MMM YYYY
ptm_date <- ptm_d %>%
  separate(`Date Collected`, into = c("d", "m", "y"), remove = F) %>%
  mutate(
    `Date Collected` = paste0(d, "-", m, "-", y), # `Date Collected` = format.Date(`Date Collected`,"%d %b %Y"),
    `Date of Extraction` = format.Date(`Date of Extraction`, "%d %b %Y")
  )
```

## Habitat and Locality
```{r countries}
# countries
ptm_country <- ptm_d %>%
  mutate(
    country_names = str_extract(Locality, "Japan|Italy|Taiwan|Chile|Australia"),
    can = if_else(str_detect(Locality, "Hakai,?|Vancouver\\sIsland|Calvert\\sIsland|Quadra\\sIsland|Bamfield"),
      "Canada,British Columbia", ""
    )
  ) %>%
  separate(can, c("coun", "prov"), ",") %>%
  mutate(
    Country = if_else(!is.na(Country), Country, country_names),
    Country = if_else(!is.na(Country), Country, coun),
    StateProvince = if_else(!is.na(StateProvince), StateProvince, prov)
  ) %>% 
  select(-coun,-country_names,-prov)

check <- select(ptm_country, `PTM#`, Country, `StateProvince`, Locality, Longitude, Latitude)
```

# Clean up names
- removed the \n (new line)
- trimmed off whitespace
```{r}
ptm_names <- ptm_country %>%
  mutate(
    `Final determination` = str_trim(`Final determination`, side = "both"),
    `Final determination` = str_remove(`Final determination`, "\\n"))

```

# Voucher Location
- gets herbarium and no voucher cases
```{r}
ptm_voucher <- ptm_names %>% 
  mutate(herb = if_else(str_detect(`Accession #`,"A"),"Herbarium","na"),
         none = if_else(str_detect(`Accession #`,"N|no voucher"),"None",""),
         `Voucher location` = if_else(is.na(`Voucher location`),herb,`Voucher location`),
         `Voucher location` = if_else(str_detect(`Voucher location`,"na"),none,`Voucher location`)) %>% 
  select(-herb,-none) %>% 
  arrange(`PTM#`)

instructions <- ptm_voucher[nrow(ptm_voucher),]
rest_ptm <- ptm_voucher[1:nrow(ptm_voucher)-1,]
ptm_complete <- rbind(instructions, rest_ptm)
```

# Save the file
```{r}
write_csv(ptm_complete, "./output/masterlist_clean.csv", na = "")
```
