==> 02_integrate_photos.Rmd <==
---
title: "02_Photos"
author: "Jasmine"
date: "12/02/2020"
output: html_document
---
## Before you start
1. connect to ubc VPN
2. connect to server

```{r libraries}
library(tidyverse)
library(ptm)
library(googlesheets4)
```

## Pick some files
- after changing this part you can hit Run > Restart R and Run All chunks
- or run one by one if you want to know what exactly this code does
```{r make some choices}
get_data_from_csv <- FALSE # if FALSE uses data from masterlist
get_photos_from_server <- TRUE # Martone Lab server access is not needed if FALSE
```


## Read in file
- change this if you are updating the masterlist
```{r choose load data}
update_or_clean_photo <- function(update = T){
  if(update){
    read_csv("output/masterlist_clean.csv",
             col_types = "iccccccDcccccccccccccccccccnnccccciiccc")
  }
  else{
    masterlist()
  }
}

ptm <- update_or_clean_photo(update = get_data_from_csv)
```

## Fetch photos from server
- **you must be connected to the server for this to work**
- takes the longest to load
- or else used cached copy `output/DATE_ptm_photos.csv`
```{r search servers}
get_photos <- function(update = F){
  if (update) {
    #need to be connected to the server to get photo information
    p <- "/Volumes/martonelab/Photos"
    
    my_files <- list.files(
      path = p,
      all.files = T,
      full.names = T,
      recursive = T
    )
    
    #clean up output
    p_file <- as.data.frame(my_files) %>%
      separate(my_files,
               into = c("folder", "m", "p", "ph", "range", "photo"),
               sep = "/",
               extra = "merge") %>%
      separate(photo, into = c("ptm", "extension"), sep = "\\.")
    
    write_csv(p_file, 
              "./output/ptm_photos.csv", 
              na = "")
  }
  else (read_csv("./output/ptm_photos.csv"))
}

ptm_photo <- get_photos(update = get_photos_from_server)
```


get the number of photos per ptm
```{r count photos}
img_clean <- ptm_photo %>% 
  filter(str_detect(ptm,"PTM"),
         range != "Quadra Bioblitz 2019") %>% 
  mutate(ptm = str_remove(ptm, "PTM"),
         num = str_extract(ptm,"[[:digit:]]{2,4}"),
         num = as.numeric(num)) %>% 
  group_by(num) %>% 
  count(name = "Photos on server NEW")
```


bring the data together
```{r join}
ptm_p <- left_join(ptm, img_clean, by = c(`PTM#` = "num")) %>% 
  select(-`Photos on server NEW`)
```


## Saving
- choose if you are continuing on to step 03 or updating to the masterlist
```{r save data - csv}
write_csv(ptm_p, "./output/masterlist_clean_photos.csv")
```

# note update sheet = "Copy of All PTM data" to all PTM Data once confirmed
```{r save data - sheets}
gs_ptm <- sheets_get("https://docs.google.com/spreadsheets/d/1vzVIT5gjQ0yCGwbAyqB4f8pltOkLfWz27TIFTxnm9hk/edit#gid=0")

save_col <- as_tibble(ptm_p$`Photos on server`, .name_repair = "minimal")
names(save_col) <- "Photos on server NEW"

sheets_edit(gs_ptm,
            save_col, 
            sheet = "All PTM data",
            range = "AK2")
```
