==> 03_integrate_sequences.Rmd <==
---
title: "Sequences"
author: "Jasmine"
date: "12/02/2020"
output: html_document
---
# Before you start - to access files:
1. connect to UBC VPN
2. connect to server

# Known issues to fix
- currently has issues with 2 digit numbers (has the 0) but cannot edit on server unless on martone lab computer
- or sequences are in folders (need to unnest)


```{r setup, include=FALSE}
library(tidyverse)
library(ptm)
library(googlesheets4)
```

## Specify some variables first
```{r}
use_data_from_csv <- FALSE # if FALSE uses data from masterlist
get_sequences_from_server <- TRUE # Martone Lab server access is not needed if FALSE (will take a while if TRUE)
```

## open files
```{r open masterlist}
update_or_clean_seq <- function(update = T){
  if(update){
    read_csv("./output/master_list_clean_photos.csv",
                 col_types = "iccccccDcccccccccccccccccccnnccccc-iiccc")
  }
  else{
    masterlist()
  }
}

ptm <- update_or_clean_seq(update = T)

# prep file to be joined
ptm_s <- ptm %>%
  mutate(ptm = paste0("PTM", `PTM#`))
```

# if we need to update the sequences run the `update_seq` function
- will now account for different folders
```{r file access}
update_seq <- function(update = T){
  if(update){
    #searching for all the sequences on the server
    seq <- list.files("/Volumes/martonelab/Sequences/0_PTM/",
                      full.names = T,
                      recursive = T) #meaning to look through all folders
    
    #seperate out file names for files NOT in folder
    seq_file <- as.data.frame(seq) %>%
      separate(seq, into = c("general_folder","rest"), # deals with the folder information
               sep = "//", extra = "merge") %>% 
      separate(rest, into = c("folder","file","fnm"), # deals with the folder information
               sep = "/", extra = "merge") %>%
      #filter(!is.na(fnm)) %>%
      separate(file, # deals with the file name
               into = c("ptm","gene","other"),
               sep = "_") %>%
      mutate(gene = str_to_lower(gene), #standardizes some information
             gene = str_remove(gene,".ab"),
             gene = str_remove(gene,"1"))
    
    #seperate out file names for files IN FOLDERS
    seq_folder <- as.data.frame(seq) %>%
      separate(seq, into = c("general_folder","rest"), # deals with the folder information
               sep = "//", extra = "merge") %>% 
      separate(rest, into = c("folder","file","fnm"), # deals with the folder information
               sep = "/", extra = "merge") %>%
      #filter(!is.na(fnm)) %>%
      separate(fnm, # deals with the file name
               into = c("ptm","gene","other"),
               sep = "_") %>%
      mutate(gene = str_to_lower(gene), #standardizes some information
             gene = str_remove(gene,".ab"),
             gene = str_remove(gene,"1"))
    
    #combining the two variations in folder naming
    seq_clean <- bind_rows(seq_folder, seq_file)
    
    #saves
    write_csv(seq_clean, "./output/server_sequence.csv")
  }
  else(read_csv(paste0("./output/server_sequence.csv")))
}
```

- this will take a while to run
```{r read sequences}
seq_clean <- update_seq(get_sequences_from_server)
```

# dealing with retrieved files
- update the below if there are new barcode primers used
- currently we have to most commonly used sequences for CO1, psbA and rbcl
```{r genes}
genes <-  c("rbclrevnew", "gwsfn", "gwsrx", "psbar2", "psbaf", "F7553", "f|F57", "R1150K")
```

inserting the F and R for all the gene names
```{r coding FR echo=TRUE}
# clean up sequence names
seq_rev <- seq_clean %>%
  mutate(gene = str_to_lower(gene)) %>% # accounts for case differences
  filter(gene %in% genes)  # selects on the sequences with primers we are interested in

# combine meta data to the files
piv <- left_join(ptm_s, seq_clean, by = "ptm") %>%
  mutate(
    gene = if_else(gene == "", "none", gene),
    gene = if_else(is.na(gene), "none", gene),
    present = 1) %>%
  pivot_wider(
    names_from = gene, values_from = present,
    values_fn = list(present = length)
  ) %>%
  distinct(`PTM#`, .keep_all = TRUE)


combine_FR <- function(forward, reverse){
  fr <- paste0(forward, reverse)
  str_remove_all(fr,"NA")
}

#putting in forward and reverse
piv_genetics <- piv %>% 
  mutate(psbar2 = if_else(psbar2 == 1, "R",""),
         psbaf = if_else(psbaf == 1, "F",""),
         p = combine_FR(psbaf,psbar2),
         gwsfn = if_else(gwsfn == 1, "F",""),
         gwsrx = if_else(gwsrx == 1, "R",""),
         g = combine_FR(gwsfn, gwsrx),
         rbclrevnew = if_else(rbclrevnew == 1, "R",""),
         f57 =  if_else(f57 == 1, "F",""),
         r = combine_FR(f57, rbclrevnew))

molecular_complete <- piv_genetics %>% 
  mutate(CO1 = if_else(is.na(CO1), g, CO1),
         psbA = if_else(is.na(psbA), p, psbA),
         rbcL = if_else(is.na(rbcL), r, rbcL))
```

# clean up and save
```{r formatting}
last_col <- grep("Project", colnames(molecular_complete))

complete_table <- molecular_complete[1:last_col]

write_csv(complete_table,"./output/masterlist_clean_sequences.csv")
```

# note update sheet = "Copy of All PTM data" to all PTM Data once confirmed
```{r save data - sheets}
save_col <- cbind(complete_table$CO1, complete_table$psbA, complete_table$rbcL) %>% 
  as_tibble()
names(save_col) <- c("CO1 NEW","PSBA NEW", "RBCL NEW")

# CURRENTLY USING THE DUP for testing purposes
gs_ptm <- sheets_get("https://docs.google.com/spreadsheets/d/1vzVIT5gjQ0yCGwbAyqB4f8pltOkLfWz27TIFTxnm9hk/edit#gid=0")

sheets_edit(gs_ptm,
            save_col, 
            sheet = "All PTM data",
            range = "AL2") #last column
```
